+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Runtime error handler.
+
 ## [0.6.0] - 2018-01-29
 ### Changed
 - `LoggerFunc` is now `ErrorLoggerFunc`.
@@ -86,6 +90,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - This package's source code, including examples and tests.
 - Go dep files.
 
+[Unreleased]: https://github.com/gbrlsnchs/httphandler/compare/v0.6.0...HEAD
 [0.6.0]: https://github.com/gbrlsnchs/httphandler/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/gbrlsnchs/httphandler/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/gbrlsnchs/httphandler/compare/v0.4.0...v0.5.0